+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+- fix: $permutator->foo(null|false|0|[]|object) allowed
+- fix: data files checked for existence; no file_get_contents to improve speed
+- change: load_mongo no longer supports csv (json only to support nesting)
+
 ## 0.1.0
 First version release:
 - feature: permutator and loader